Add NUM_CHANNELS envvar to define the kernels instantiations to build
The corresponding macro is GSPLAT_NUM_CHANNELS. By default
(if nothing is defined), compile all number of channels that was
previously supported.

Tests only run with the needed subset of NUM_CHANNELS.
Also run tests with VERBOSE=1 by default.

Had to bump pytest to 7.1.3 due to the new dependency on pytest-env